tests: update e2e tests architecture
